refactor: 불필요한 CLI 코드 제거
