--- src/View/CLI/index.js
@@ -1,3 +1,0 @@
-export * from './InputView';
-export * from './OutputView';
-export * from './View';
