--- src/View/OutputView.js
@@ -1,7 +1,0 @@
-import { Console } from '../utils';
-
-export const OutputView = {
-  print(...message) {
-    Console.print(...message);
-  },
-};
