--- src/View/Web/index.js
@@ -1,3 +1,0 @@
-export * from './Element';
-export * from './OutputViewWeb';
-export * from './ViewWeb';
