+++ src/View/index.js
@@ -1,3 +1,3 @@
-export * from './InputView';
-export * from './OutputView';
-export * from './View';
+export * from './Element';
+export * from './OutputViewWeb';
+export * from './ViewWeb';
