+++ src/utils/index.js
@@ -1,3 +1,2 @@
-export * from './Console';
 export * from './getRandomIntInRange';
 export * from './lottoPrize';
